requirements.txt
streamlink
numpy
opencv-python
google-cloud-vision
irc
pil